Externalized widgets bar from WidgetsTray plugin
